Allow electric cartage if cheap, max temp, tweaks
Allow the usage of the electric cartage (by setting the real outdoor
temperature) if the price is below 100 öre.

Turn off heating if it's above 21.5C indoor

Do not fully turn of the heat pump if it's below freezing outdoor. I'm
not sure if the cirulation pump is turned off or not, I will play safe
for now.